+++ src/wizardPages/Page_ReviewAndRun.java
@@ -69,7 +69,7 @@ import objects.SensorSetting;
 import objects.TecplotNode;
 import results.ResultPrinter;
 import utilities.Constants;
-import utilities.Point3d;
+import utilities.Point3f;
 import utilities.Point3i;
 import wizardPages.DREAMWizard.STORMData;
 
@@ -816,7 +816,7 @@ public class Page_ReviewAndRun extends WizardPage implements AbstractWizardPage 
 					String[] parts = individualSensor.split(":");
 					if(parts.length == 3) {
 						int nodeNumber = Integer.parseInt(parts[0].trim());
-						Point3d xyz = data.getSet().getNodeStructure().getXYZEdgeFromIJK(data.getSet().getNodeStructure().getIJKFromNodeNumber(nodeNumber));
+						Point3f xyz = data.getSet().getNodeStructure().getXYZEdgeFromIJK(data.getSet().getNodeStructure().getIJKFromNodeNumber(nodeNumber));
 						nodesToReplace.put(parts[0], xyz.toString());
 					} 
 				}
